rod/requirements.txt
Flask-Cors
Flask-Restless
Flask-Login
SQLAlchemy
